+++ src/auth/auth.controller.ts
@@ -1,0 +1,7 @@
+import { Controller } from '@nestjs/common';
+import { AuthService } from './auth.service';
+
+@Controller('auth')
+export class AuthController {
+  constructor(private readonly authService: AuthService) {}
+}
