+++ src/auth/strategyies/jwt.refresh.ts
@@ -7,7 +7,7 @@ import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { Request } from 'express';
 import { env } from 'src/common/configs/env';
-import { TokenPayloadDto } from '../dto/tokenPayload';
+import { TokenPayloadDto } from '../../common/dto/tokenPayload';
 import { MembersService } from 'src/members/members.service';
 interface JwtRefreshPayload extends TokenPayloadDto {
   refreshToken: string;
