+++ src/auth/strategyies/jwt.strategy.ts
@@ -3,13 +3,10 @@ import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { Request } from 'express';
 import { env } from 'src/common/configs/env';
-import { TokenPayloadDto } from '../dto/tokenPayload';
+import { TokenPayloadDto } from '../../common/dto/tokenPayload';
 
 @Injectable()
-export class JwtOptionalStrategy extends PassportStrategy(
-  Strategy,
-  'optional',
-) {
+export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
   constructor() {
     super({
       jwtFromRequest: ExtractJwt.fromExtractors([
@@ -26,7 +23,16 @@ export class JwtOptionalStrategy extends PassportStrategy(
     });
   }
 
-  async validate(payload: TokenPayloadDto | null) {
-    return payload ? payload : null;
+  async validate(payload: any) {
+    if (!payload || !payload.id) {
+      throw new UnauthorizedException('Invalid token payload');
+    }
+
+    const user: TokenPayloadDto = {
+      id: payload.id,
+      email: payload.email,
+    };
+
+    return user;
   }
 }
